refactor: define function to create externalId data files
